+++ Example/QRCodeReader.swift/ViewController.swift
@@ -1,28 +1,28 @@
 /*
- * QRCodeReader.swift
- *
- * Copyright 2014-present Yannick Loriot.
- * http://yannickloriot.com
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in
- * all copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
- * THE SOFTWARE.
- *
- */
+* QRCodeReader.swift
+*
+* Copyright 2014-present Yannick Loriot.
+* http://yannickloriot.com
+*
+* Permission is hereby granted, free of charge, to any person obtaining a copy
+* of this software and associated documentation files (the "Software"), to deal
+* in the Software without restriction, including without limitation the rights
+* to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+* copies of the Software, and to permit persons to whom the Software is
+* furnished to do so, subject to the following conditions:
+*
+* The above copyright notice and this permission notice shall be included in
+* all copies or substantial portions of the Software.
+*
+* THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+* IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+* FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+* AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+* LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+* OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+* THE SOFTWARE.
+*
+*/
 
 import UIKit
 import AVFoundation
@@ -36,7 +36,9 @@ class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
       reader.delegate               = self
 
       reader.completionBlock = { (result: QRCodeReaderResult?) in
-        print("Completion with result: \(result!.value) of type \(result!.type)")
+        if let result = result {
+          print("Completion with result: \(result.value) of type \(result.metadataType)")
+        }
       }
 
       presentViewController(reader, animated: true, completion: nil)
@@ -52,11 +54,15 @@ class ViewController: UIViewController, QRCodeReaderViewControllerDelegate {
   // MARK: - QRCodeReader Delegate Methods
 
   func reader(reader: QRCodeReaderViewController, didScanResult result: QRCodeReaderResult) {
-    self.dismissViewControllerAnimated(true, completion: { [unowned self] () -> Void in
-        let alert = UIAlertController(title: "QRCodeReader", message: String (format:"%@ (of type %@)", result.value!, result.type!), preferredStyle: .Alert)
+    self.dismissViewControllerAnimated(true, completion: { [weak self] in
+      let alert = UIAlertController(
+        title: "QRCodeReader",
+        message: String (format:"%@ (of type %@)", result.value, result.metadataType),
+        preferredStyle: .Alert
+      )
       alert.addAction(UIAlertAction(title: "OK", style: .Cancel, handler: nil))
 
-      self.presentViewController(alert, animated: true, completion: nil)
+      self?.presentViewController(alert, animated: true, completion: nil)
       })
   }